+++ examples/testing/go.mod
@@ -1,5 +1,5 @@
-module main
+module test
 
 go 1.16
 
-
+require github.com/fisik-yum/oddmap v0.1.0 // indirect
